HTTP Server: Working version
There's a behaviour that I cannot explain yet, it might be due to how
Chrome makes the requests. After the number of jobs specified in take(N)
are finished, I still need to refresh the browser once more to get the
worker that got the last job to exit. EDIT: CONFIRMED! This doesn't
happen when calling with curl.